Revert "Sync system.error_log and system.errors"
