+++ src/Interpreters/ErrorLog.cpp
@@ -10,8 +10,8 @@
 #include <Interpreters/ErrorLog.h>
 #include <Parsers/ExpressionElementParsers.h>
 #include <Parsers/parseQuery.h>
-#include <Common/SymbolsHelper.h>
 
+#include <vector>
 
 namespace DB
 {
@@ -19,7 +19,6 @@ namespace DB
 ColumnsDescription ErrorLogElement::getColumnsDescription()
 {
     ParserCodec codec_parser;
-    DataTypePtr symbolized_type = std::make_shared<DataTypeArray>(std::make_shared<DataTypeLowCardinality>(std::make_shared<DataTypeString>()));
     return ColumnsDescription {
         {
                 "hostname",
@@ -52,18 +51,6 @@ ColumnsDescription ErrorLogElement::getColumnsDescription()
                 "Error name."
             },
         {
-                "last_error_time",
-                std::make_shared<DataTypeDateTime>(),
-                parseQuery(codec_parser, "(ZSTD(1))", 0, DBMS_DEFAULT_MAX_PARSER_DEPTH, DBMS_DEFAULT_MAX_PARSER_BACKTRACKS),
-                "The time when the last error happened."
-            },
-        {
-                "last_error_message",
-                std::make_shared<DataTypeString>(),
-                parseQuery(codec_parser, "(ZSTD(1))", 0, DBMS_DEFAULT_MAX_PARSER_DEPTH, DBMS_DEFAULT_MAX_PARSER_BACKTRACKS),
-                "Message for the last error."
-            },
-        {
                 "value",
                 std::make_shared<DataTypeUInt64>(),
                 parseQuery(codec_parser, "(ZSTD(3))", 0, DBMS_DEFAULT_MAX_PARSER_DEPTH, DBMS_DEFAULT_MAX_PARSER_BACKTRACKS),
@@ -74,30 +61,6 @@ ColumnsDescription ErrorLogElement::getColumnsDescription()
                 std::make_shared<DataTypeUInt8>(),
                 parseQuery(codec_parser, "(ZSTD(1))", 0, DBMS_DEFAULT_MAX_PARSER_DEPTH, DBMS_DEFAULT_MAX_PARSER_BACKTRACKS),
                 "Remote exception (i.e. received during one of the distributed queries)."
-            },
-        {
-                "last_error_query_id",
-                std::make_shared<DataTypeString>(),
-                parseQuery(codec_parser, "(ZSTD(1))", 0, DBMS_DEFAULT_MAX_PARSER_DEPTH, DBMS_DEFAULT_MAX_PARSER_BACKTRACKS),
-                "Id of a query that caused the last error (if available)."
-            },
-        {
-                "last_error_trace",
-                std::make_shared<DataTypeArray>(std::make_shared<DataTypeUInt64>()),
-                parseQuery(codec_parser, "(ZSTD(1))", 0, DBMS_DEFAULT_MAX_PARSER_DEPTH, DBMS_DEFAULT_MAX_PARSER_BACKTRACKS),
-                "A stack trace that represents a list of physical addresses where the called methods are stored."
-            },
-        {
-                "symbols",
-                symbolized_type,
-                parseQuery(codec_parser, "(ZSTD(1))", 0, DBMS_DEFAULT_MAX_PARSER_DEPTH, DBMS_DEFAULT_MAX_PARSER_BACKTRACKS),
-                "If the symbolization is enabled, contains demangled symbol names, corresponding to the `trace`."
-            },
-        {
-                "lines",
-                symbolized_type,
-                parseQuery(codec_parser, "(ZSTD(1))", 0, DBMS_DEFAULT_MAX_PARSER_DEPTH, DBMS_DEFAULT_MAX_PARSER_BACKTRACKS),
-                "If the symbolization is enabled, contains strings with file names with line numbers, corresponding to the `trace`."
             }
     };
 }
@@ -111,33 +74,8 @@ void ErrorLogElement::appendToBlock(MutableColumns & columns) const
     columns[column_idx++]->insert(event_time);
     columns[column_idx++]->insert(code);
     columns[column_idx++]->insert(ErrorCodes::getName(code));
-    columns[column_idx++]->insert(error_time_ms / 1000);
-    columns[column_idx++]->insert(error_message);
     columns[column_idx++]->insert(value);
     columns[column_idx++]->insert(remote);
-    columns[column_idx++]->insert(query_id);
-
-    std::vector<UInt64> error_trace_array;
-    error_trace_array.reserve(error_trace.size());
-    for (auto * ptr : error_trace)
-        error_trace_array.emplace_back(reinterpret_cast<UInt64>(ptr));
-
-    columns[column_idx++]->insert(Array(error_trace_array.begin(), error_trace_array.end()));
-
-#if defined(__ELF__) && !defined(OS_FREEBSD)
-    if (symbolize)
-    {
-        auto [symbols, lines] = generateArraysSymbolsLines(error_trace_array);
-
-        columns[column_idx++]->insert(symbols);
-        columns[column_idx++]->insert(lines);
-    }
-    else
-#endif
-    {
-        columns[column_idx++]->insertDefault();
-        columns[column_idx++]->insertDefault();
-    }
 }
 
 struct ValuePair
@@ -161,13 +99,8 @@ void ErrorLog::stepFunction(TimePoint current_time)
             ErrorLogElement local_elem {
                 .event_time=event_time,
                 .code=code,
-                .error_time_ms=error.local.error_time_ms,
-                .error_message=error.local.message,
                 .value=error.local.count - previous_values.at(code).local,
-                .remote=false,
-                .query_id=error.local.query_id,
-                .error_trace=error.local.trace,
-                .symbolize=symbolize
+                .remote=false
             };
             this->add(std::move(local_elem));
             previous_values[code].local = error.local.count;
@@ -177,13 +110,8 @@ void ErrorLog::stepFunction(TimePoint current_time)
             ErrorLogElement remote_elem {
                 .event_time=event_time,
                 .code=code,
-                .error_time_ms=error.remote.error_time_ms,
-                .error_message=error.remote.message,
                 .value=error.remote.count - previous_values.at(code).remote,
-                .remote=true,
-                .query_id=error.remote.query_id,
-                .error_trace=error.remote.trace,
-                .symbolize=symbolize
+                .remote=true
             };
             add(std::move(remote_elem));
             previous_values[code].remote = error.remote.count;
