+++ src/Storages/System/StorageSystemErrors.cpp
@@ -6,8 +6,6 @@
 #include <Interpreters/Context.h>
 #include <Common/ErrorCodes.h>
 #include <Core/Settings.h>
-#include <Common/SymbolsHelper.h>
-#include <DataTypes/DataTypeLowCardinality.h>
 
 
 namespace DB
@@ -19,7 +17,6 @@ namespace Setting
 
 ColumnsDescription StorageSystemErrors::getColumnsDescription()
 {
-    DataTypePtr symbolized_type = std::make_shared<DataTypeArray>(std::make_shared<DataTypeLowCardinality>(std::make_shared<DataTypeString>()));
     return ColumnsDescription
     {
         { "name",                    std::make_shared<DataTypeString>(), "Name of the error (errorCodeToName)."},
@@ -29,9 +26,7 @@ ColumnsDescription StorageSystemErrors::getColumnsDescription()
         { "last_error_message",      std::make_shared<DataTypeString>(), "Message for the last error."},
         { "last_error_trace",        std::make_shared<DataTypeArray>(std::make_shared<DataTypeUInt64>()), "A stack trace that represents a list of physical addresses where the called methods are stored."},
         { "remote",                  std::make_shared<DataTypeUInt8>(), "Remote exception (i.e. received during one of the distributed queries)."},
-        { "query_id",                std::make_shared<DataTypeString>(), "Id of a query that caused the last error (if available)." },
-        { "last_error_symbols",      symbolized_type, "Contains demangled symbol names, corresponding to the `trace`." },
-        { "last_error_lines",        symbolized_type, "Contains strings with file names with line numbers, corresponding to the `trace`." },
+        { "query_id",                std::make_shared<DataTypeString>(), "Id of a query that caused an error (if available)." },
     };
 }
 
@@ -48,26 +43,16 @@ void StorageSystemErrors::fillData(MutableColumns & res_columns, ContextPtr cont
             res_columns[col_num++]->insert(error.count);
             res_columns[col_num++]->insert(error.error_time_ms / 1000);
             res_columns[col_num++]->insert(error.message);
+            {
+                Array trace_array;
+                trace_array.reserve(error.trace.size());
+                for (size_t i = 0; i < error.trace.size(); ++i)
+                    trace_array.emplace_back(reinterpret_cast<intptr_t>(error.trace[i]));
 
-            std::vector<UInt64> error_trace_uint;
-            error_trace_uint.reserve(error.trace.size());
-            for (size_t i = 0; i < error.trace.size(); ++i)
-                error_trace_uint.emplace_back(reinterpret_cast<UInt64>(error.trace[i]));
-
-            res_columns[col_num++]->insert(Array(error_trace_uint.begin(), error_trace_uint.end()));
+                res_columns[col_num++]->insert(trace_array);
+            }
             res_columns[col_num++]->insert(remote);
             res_columns[col_num++]->insert(error.query_id);
-
-
-#if defined(__ELF__) && !defined(OS_FREEBSD)
-            auto [symbols, lines] = generateArraysSymbolsLines(error_trace_uint);
-
-            res_columns[col_num++]->insert(symbols);
-            res_columns[col_num++]->insert(lines);
-#else
-            res_columns[col_num++]->insertDefault();
-            res_columns[col_num++]->insertDefault();
-#endif
         }
     };
