+++ tests/queries/0_stateless/02117_show_create_table_system.reference
@@ -265,9 +265,7 @@ CREATE TABLE system.errors
     `last_error_message` String,
     `last_error_trace` Array(UInt64),
     `remote` UInt8,
-    `query_id` String,
-    `last_error_symbols` Array(LowCardinality(String)),
-    `last_error_lines` Array(LowCardinality(String))
+    `query_id` String
 )
 ENGINE = SystemErrors
 COMMENT 'Contains a list of all errors which have ever happened including the error code, last time and message with unsymbolized stacktrace.'
