+++ tests/queries/0_stateless/03172_error_log_table_not_empty.reference
@@ -4,6 +4,3 @@
 1
 1
 1
-1
-1
-1
